Update namespace Xamarin Core
* Update file and folder name from Com.OneSignal.Core to OneSignalSDK.Xamarin.Core
* Update namespace from Com.OneSignal.Core to OneSignalSDK.Xamarin.Core
* Update file name and namespace of OneSignalSDK to OneSignalSDKInternal
