+++ OneSignalSDK.Xamarin.Core/EmailSubscriptionState.cs
@@ -1,12 +1,12 @@
 ﻿using System;
-namespace Com.OneSignal.Core {
+namespace OneSignalSDK.Xamarin.Core {
    [Serializable]
-   public sealed class PushSubscriptionState {
+   public sealed class EmailSubscriptionState {
       /// <summary>
       /// Unique id of this subscription
       /// </summary>
       /// <remarks>See https://documentation.onesignal.com/docs/users#player-id for more information</remarks>
-      public string userId;
+      public string emailUserId;
 
       /// <summary>
       /// Whether this subscription is currently active
@@ -14,10 +14,8 @@ namespace Com.OneSignal.Core {
       public bool isSubscribed;
 
       /// <summary>
-      /// The unique token provided by the device's operating system used to send push notifications
+      /// Email address of the user which this subscription was created for
       /// </summary>
-      public string pushToken;
-
-      public bool isPushDisabled;
+      public string emailAddress;
    }
 }
