+++ OneSignalSDK.Xamarin.Core/InAppMessageOutcome.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Com.OneSignal.Core {
+namespace OneSignalSDK.Xamarin.Core {
    public class InAppMessageOutcome {
       public string name;
       public float weight;
