+++ OneSignalSDK.Xamarin.Core/NotificationOpenedResult.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-namespace Com.OneSignal.Core {
+namespace OneSignalSDK.Xamarin.Core {
    [Serializable]
    public sealed class NotificationOpenedResult {
       public NotificationAction action;
